finall refactoring

+++ conftest.py
@@ -1,9 +1,6 @@
 from selenium import webdriver
 from webdriver_manager.chrome import ChromeDriverManager
-from orm.client import MySQLClient
-import subprocess
-from time import time, sleep
-import requests
+from local_code.orm.client import MySQLClient
 import pytest
 import logging
 import shutil
@@ -12,10 +9,11 @@ import os
 
 
 def pytest_addoption(parser):
-    parser.addoption("--url", default="http://0.0.0.0:8080")
+    parser.addoption("--url", default="http://0.0.0.0:8081/")
     parser.addoption("--mock", default="http://172.18.0.3:5000/")
     parser.addoption("--debug_log", action='store_true')
     parser.addoption("--mysql", default="127.0.0.1:3306")
+    parser.addoption("--selenoid", action='store_true')
 
 
 @pytest.fixture(scope='session')
@@ -24,7 +22,8 @@ def config(request):
     mock = request.config.getoption("--mock")
     mysql = request.config.getoption("--mysql")
     debug_log = request.config.getoption("--debug_log")
-    return { 'url': url, 'mock': mock, 'mysql': mysql, 'debug_log': debug_log}
+    selenoid = "http://127.0.0.1:4444/wd/hub" if request.config.getoption('--selenoid') else None
+    return { 'url': url, 'mock': mock, 'mysql': mysql, 'debug_log': debug_log, "selenoid": selenoid }
 
 
 def pytest_configure(config):
@@ -32,10 +31,11 @@ def pytest_configure(config):
     mysql_client.connect()
     config.mysql_client = mysql_client
     test_dir = os.path.join('tmp', 'test')
-    if os.path.exists(test_dir):
-        shutil.rmtree(test_dir)
-    os.makedirs(test_dir)
-    config.test_dir = test_dir
+    if not hasattr(config, 'workerinput'):
+        if os.path.exists(test_dir):
+            shutil.rmtree(test_dir)
+        os.makedirs(test_dir)
+    config.base_test_dir = test_dir
     # timeout = 60
     # start_time = time()
     # subprocess.Popen(["sudo", "docker-compose", "up"])
@@ -60,13 +60,32 @@ def mysql_client(request):
 
 
 @pytest.fixture(scope='function')
-def driver(config):
-    os.environ['WDM_LOG_LEVEL'] = '0'
-    manager = ChromeDriverManager(version='latest')
-    driver = webdriver.Chrome(executable_path=manager.install())
-    driver.get(config['url'])
-    yield driver
-    driver.quit()
+def driver(config, get_driver):
+    get_driver.get(config['url'])
+    yield get_driver
+    get_driver.quit()
+
+
+@pytest.fixture(scope='function')
+def get_driver(config):
+    if config["selenoid"] is not None:
+        capabilities = {
+            "browserName": "chrome",
+            "browserVersion": "97.0",
+            # "applicationContainers": [f"{config['url']}:myapp"],
+            "selenoid:options": {
+                "enableVNC": True,
+                "enableVideo": False
+            }
+        }
+        browser = webdriver.Remote(command_executor=config["selenoid"], desired_capabilities=capabilities)
+    else:
+        os.environ['WDM_LOG_LEVEL'] = '0'
+        manager = ChromeDriverManager(version='latest')
+        browser = webdriver.Chrome(executable_path=manager.install())
+
+    browser.maximize_window()
+    return browser
 
 
 repo_root = os.path.dirname(__file__)
@@ -76,13 +95,13 @@ repo_root = os.path.dirname(__file__)
 def test_dir(request, config):
     test_name = request._pyfuncitem.nodeid.replace(':', '_').replace('/', '_')
     # test_name = request.node.name
-    dir_for_test = os.path.join(request.config.test_dir, test_name)
+    dir_for_test = os.path.join(request.config.base_test_dir, test_name)
     os.makedirs(dir_for_test)
     return dir_for_test
 
 
 @pytest.fixture(scope='function')
-def logger(config, test_dir): # Поиграться и разобраться с тем как это работает
+def logger(config, test_dir):
     log_formatter = logging.Formatter('%(asctime)s - %(filename)s - %(levelname)s: %(message)s')
     log_file = os.path.join(test_dir, 'test.log')
     log_level = logging.DEBUG if config['debug_log'] else logging.INFO
